pyproject.toml
[tool.poetry]
name = "interactive-live2d-voice"
version = "0.1.0"
description = "Interactive chat interface with Live2D character and voice synthesis"
authors = ["faith6 <root39293@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
jinja2 = "^3.0.1"
python-dotenv = "^0.19.0"
openai = "^0.27.0"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"